+++ olive/data/component/hf/text_classification.py
@@ -1,0 +1,105 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+# --------------------------------------------------------------------------
+from typing import List
+
+from olive.common.config_utils import validate_config
+from olive.data.component.dataset import BaseDataset
+from olive.data.component.hf.utils import TokenizerConfig, map_pre_process, truncate_dataset
+
+
+class CommonClassificationParams(TokenizerConfig):
+    """Common parameters for text classification and ner tasks"""
+
+    input_cols: List[str]  # input columns
+    label_cols: List[str]  # label columns, we take in a list but only use the first one for now
+    max_samples: int = None  # max number of samples to use, None for all
+    max_length: int = None  # max length of input sequence, used for padding and truncation
+    is_split_into_words: bool = False  # whether the input is already split into words
+
+
+class TextClassificationParams(CommonClassificationParams):
+    """Parameters for text classification task"""
+
+    align_labels: bool = False  # whether to align labels with tokens
+    model_config_path: str = None  # path to model config file, used to align labels. If None, use model_name instead
+
+
+def text_classification_pre_process(dataset, model_name, **kwargs):
+    """Pre-process data for text classification task."""
+    from transformers import AutoConfig
+
+    args = validate_config(kwargs, TextClassificationParams, warn_unused_keys=True)
+
+    # do this before tokenization to avoid unnecessary tokenization
+    dataset = truncate_dataset(dataset, args.max_samples)
+
+    tokenizer, tokenization_kwargs = args.get_tokenizer(model_name, args.max_length)
+    tokenization_kwargs["is_split_into_words"] = args.is_split_into_words
+
+    # TODO: add the complete data operation mapping like:
+    # align_labels -> align_labels_with_mapping
+    # Also to support customized operation arguments from users
+    if args.align_labels:
+        model_hf_config = AutoConfig.from_pretrained(args.model_config_path or model_name)
+        if model_hf_config and model_hf_config.label2id:
+            dataset = dataset.align_labels_with_mapping(model_hf_config.label2id, args.label_cols[0])
+
+    def tokenize_and_add_labels(examples):
+        tokenized_examples = tokenizer(examples, **tokenization_kwargs)
+        # TODO: support multiple label columns if needed
+        tokenized_examples["label"] = examples[args.label_cols[0]]
+        return tokenized_examples
+
+    tokenized_dataset = map_pre_process(dataset, tokenize_and_add_labels, batched=True, remove_cols=True)
+    # label_cols is ["label"] since we added label_cols[0] as "label" to tokenized_inputs
+    return BaseDataset(tokenized_dataset, label_cols=["label"], max_samples=args.max_samples)
+
+
+def ner_pre_process(dataset, model_name, **kwargs):
+    """Pre-process data for ner task."""
+
+    args = validate_config(kwargs, CommonClassificationParams, warn_unused_keys=True)
+
+    # do this before tokenization to avoid unnecessary tokenization
+    dataset = truncate_dataset(dataset, args.max_samples)
+
+    tokenizer, tokenization_kwargs = args.get_tokenizer(model_name, args.max_length)
+    tokenization_kwargs["is_split_into_words"] = args.is_split_into_words
+
+    def align_labels_with_tokens(labels, word_ids):
+        new_labels = []
+        current_word = None
+        for word_id in word_ids:
+            if word_id != current_word:
+                # Start of a new word!
+                current_word = word_id
+                label = 0 if word_id is None else labels[word_id]
+                new_labels.append(label)
+            elif word_id is None:
+                # Special token
+                new_labels.append(0)
+            else:
+                # Same word as previous token
+                label = labels[word_id]
+                # If the label is B-XXX we change it to I-XXX
+                if label % 2 == 1:
+                    label += 1
+                new_labels.append(label)
+        return new_labels
+
+    def tokenize_and_align_labels(examples):
+        tokenized_examples = tokenizer(examples, **tokenization_kwargs)
+        all_labels = examples[args.label_cols[0]]
+        new_labels = []
+        for i, labels in enumerate(all_labels):
+            word_ids = tokenized_examples.word_ids(i)
+            new_labels.append(align_labels_with_tokens(labels, word_ids))
+
+        tokenized_examples["label"] = new_labels
+        return tokenized_examples
+
+    tokenized_dataset = map_pre_process(dataset, tokenize_and_align_labels, batched=True, remove_cols=True)
+    # label_cols is ["label"] since we added label_cols[0] as "label" to tokenized_inputs
+    return BaseDataset(tokenized_dataset, label_cols=["label"], max_samples=args.max_samples)
