+++ olive/data/component/pre_process_data.py
@@ -6,13 +6,12 @@
 
 from copy import deepcopy
 
-from olive.data.component.dataset import BaseDataset
-from olive.data.component.text_generation import (
+from olive.data.component.hf.text_classification import ner_pre_process, text_classification_pre_process
+from olive.data.component.hf.text_generation import (
     TextGenDatasetType,
     text_gen_corpus_pre_process,
     text_gen_pair_pre_process,
 )
-from olive.data.component.utils import map_pre_process
 from olive.data.registry import Registry
 
 
@@ -35,90 +34,44 @@ def huggingface_pre_process(_dataset, model_name, input_cols, label_cols, max_sa
     """Pre-process data.
 
     Args:
-        data (object): Data to be pre-processed.
+        _dataset (object): Data to be pre-processed.
+        model_name (str): Name of the huggingface model.
+        input_cols (List[str]): Input columns.
+        label_cols (List[str]): Label columns.
         **kwargs: Additional arguments.
+            The arguments are the fields in olive.data.component.hf.text_classification.TextClassificationParams.
 
     Returns:
         object: Pre-processed data.
     """
-    from transformers import AutoConfig, AutoTokenizer
 
-    def _tokenizer_and_align_labels(examples):
-        tokenizer = AutoTokenizer.from_pretrained(model_name, padding_side=kwargs.get("padding_side", "right"))
-        tokenized_inputs = tokenizer(
-            *[examples[input_col] for input_col in input_cols],
-            padding=kwargs.get("padding", True),
-            truncation=kwargs.get("truncation", True),
-            is_split_into_words=kwargs.get("is_split_into_words", False),
-            add_special_tokens=kwargs.get("add_special_tokens", True),
-        )
-        # TODO: support multiple label columns if needed
-        tokenized_inputs["label"] = examples[label_cols[0]]
-        # huggingface dataset api limit to return dict and arrow table
-        return tokenized_inputs
+    all_kwargs = deepcopy(kwargs)
+    all_kwargs.update({"max_samples": max_samples, "input_cols": input_cols, "label_cols": label_cols})
 
-    model_config_path = kwargs.pop("model_config_path", None)
-    # TODO: add the complete data operation mapping like:
-    # align_labels -> align_labels_with_mapping
-    # Also to support customized operation arguments from users
-    if kwargs.pop("align_labels", False):
-        model_hf_config = AutoConfig.from_pretrained(model_config_path or model_name)
-        if model_hf_config and model_hf_config.label2id:
-            _dataset = _dataset.align_labels_with_mapping(model_hf_config.label2id, label_cols[0])
-
-    tokenized_datasets = map_pre_process(_dataset, _tokenizer_and_align_labels, batched=True, remove_cols=True)
-    # label_cols is ["label"] since we added label_cols[0] as "label" to tokenized_inputs
-    return BaseDataset(tokenized_datasets, label_cols=["label"], max_samples=max_samples)
+    return text_classification_pre_process(_dataset, model_name, **all_kwargs)
 
 
 @Registry.register_pre_process()
 def ner_huggingface_preprocess(_dataset, model_name, input_cols, label_cols, max_samples=None, **kwargs):
     """
     Pre-process data for ner task.
+
+    Args:
+        _dataset (object): Data to be pre-processed.
+        model_name (str): Name of the huggingface model.
+        input_cols (List[str]): Input columns.
+        label_cols (List[str]): Label columns.
+        max_samples (int, optional): Max number of samples to use. Defaults to None.
+        **kwargs: Additional arguments.
+            The arguments are the fields in olive.data.component.hf.text_classification.CommonClassificationParams.
+
+    Returns:
+        object: Pre-processed data.
     """
-    from transformers import AutoTokenizer
+    all_kwargs = deepcopy(kwargs)
+    all_kwargs.update({"max_samples": max_samples, "input_cols": input_cols, "label_cols": label_cols})
 
-    def _align_labels_with_tokens(labels, word_ids):
-        new_labels = []
-        current_word = None
-        for word_id in word_ids:
-            if word_id != current_word:
-                # Start of a new word!
-                current_word = word_id
-                label = 0 if word_id is None else labels[word_id]
-                new_labels.append(label)
-            elif word_id is None:
-                # Special token
-                new_labels.append(0)
-            else:
-                # Same word as previous token
-                label = labels[word_id]
-                # If the label is B-XXX we change it to I-XXX
-                if label % 2 == 1:
-                    label += 1
-                new_labels.append(label)
-        return new_labels
-
-    def _tokenizer_and_align_labels(examples):
-        tokenizer = AutoTokenizer.from_pretrained(model_name, padding_side=kwargs.get("padding_side", "right"))
-        tokenized_inputs = tokenizer(
-            *[examples[input_col] for input_col in input_cols],
-            padding=kwargs.get("padding", True),
-            truncation=kwargs.get("truncation", True),
-            is_split_into_words=kwargs.get("is_split_into_words", True),
-            add_special_tokens=kwargs.get("add_special_tokens", False),
-        )
-        all_labels = examples[label_cols[0]]
-        new_labels = []
-        for i, labels in enumerate(all_labels):
-            word_ids = tokenized_inputs.word_ids(i)
-            new_labels.append(_align_labels_with_tokens(labels, word_ids))
-
-        tokenized_inputs["label"] = new_labels
-        return tokenized_inputs
-
-    tokenized_datasets = map_pre_process(_dataset, _tokenizer_and_align_labels, batched=True, remove_cols=True)
-    return BaseDataset(tokenized_datasets, label_cols=["label"], max_samples=max_samples)
+    return ner_pre_process(_dataset, model_name, **all_kwargs)
 
 
 @Registry.register_pre_process()
